IN-111 refactor: split the logic into useHandleCalendar and useFetchLocations
